\name{addlines}
\alias{addlines}

\title{Adds lines between nodes to an existing plot of nodes}

\description{
Adds lines between nodes to an existing plot of nodes
}

\usage{
addlines(U, Y, col = "green", lwd = 1, lty = 1)
}

\arguments{
  \item{U}{an n x 2 matrix of node locations }
  \item{Y}{a symmetric matrix}
  \item{col}{color of the lines}
  \item{lwd}{width of the lines}
  \item{lty}{line type}
}

\value{
NULL
}

\author{Peter Hoff}


\keyword{multivariate}% at least one, from doc/KEYWORDS
