\name{eigenmodel_setup}
\alias{eigenmodel_setup}

\title{Setup constants and starting values for an eigenmodel fit}
\description{
Setup constants and starting values for an eigenmodel fit
}
\usage{
eigenmodel_setup(R = 0, seed = 1, em_env = .GlobalEnv)
}

\arguments{
  \item{R}{non-negative integer rank of the approximating matrix  }
  \item{seed}{a random seed }
  \item{em_env}{enviromnemt within which to do the fitting}
}
\value{NULL}

\author{Peter Hoff} 

\keyword{ multivariate }
\keyword{ models }

