\name{plot.eigenmodel_post}
\alias{plot.eigenmodel_post}
\title{Plot the output of an eigenmodel fit}
\description{
A graphical display of MCMC output and posterior estimates 
of model parameters in an eigenmodel fit. Includes 95 percent 
quantile-based posterior 
confidence intervals of regression coefficients. 
}
\usage{
\method{plot}{eigenmodel_post}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an object of class \code{eigenmodel_post}  }
  \item{...}{additional plotting options  }
}

\value{NULL}
\author{Peter Hoff}

\keyword{multivariate}% at least one, from doc/KEYWORDS
