\name{ULU}

\alias{ULU}

\title{Computes a matrix from its eigenvalue decomposition}

\description{
Computes a matrix from its eigenvalue decomposition
}
\usage{
ULU(UL)
}

\arguments{
  \item{UL}{ a list with first component ``U'', an n x r matrix and the 
      second component ``L'' an r x r diagonal matrix   }
             }

\value{
an n x n matrix }
\author{ Peter Hoff }

\keyword{ multivariate }
\keyword{ models }

