\name{XB}

\alias{XB}

\title{Computes a sociomatrix of  regression effects }

\description{
Computes a sociomatrix of  regression effects }

\usage{
XB(X, b)
}

\arguments{
  \item{X}{ an n x n x p array  }
  \item{b}{ a p x 1 vector  }
}

\value{
 an n x n matrix 
}

\author{ Peter Hoff }

\keyword{ multivariate }
\keyword{ models }

