\name{Y_Gen}
\alias{Y_Gen}
\docType{data}
\title{ Relations between words in the 1st chapter of Genesis  }
\description{
The i,j th entry of this matrix is the numerical count of the 
number of times word i was next to word j in the first chapter 
of Genesis. 
}
\usage{data(Y_Gen)}
\examples{
data(Y_Gen)
}
\keyword{datasets}
