\name{Y_impute}
\alias{Y_impute}

\title{ Impute missing values of a sociomatrix }
\description{
 Impute missing values of a sociomatrix 
}
\usage{
Y_impute()
}

\details{
Imputes missing values of a sociomatrix from a matrix of 
latent variables and an  ordered-probit specification.
}

\value{
 symmetric matrix
}

\author{Peter Hoff  }

\keyword{ multivariate }
\keyword{ models }

