\name{Y_Pro}
\alias{Y_Pro}
\docType{data}
\title{ Butland's protein-protein interaction data}
\description{
 Butland's protein-protein interaction data
}
\usage{data(Y_Pro)}

\references{Butland et al (2005)  ``Interaction network containing conserved and essential protein complexes in Escherichia coli''}

\examples{
data(Y_Pro)
}

\keyword{datasets}
