\name{YX_Friend}
\alias{YX_Friend}
\docType{data}
\title{Sex, race and friendship data from a 12th grade classroom }
\description{
A list in which \code{Y} encodes the presence of a friendship tie 
between 90 12th graders. The array \code{X} indicates pairs of the 
same sex and of the same race. 
}
\usage{data(YX_Friend)}
\source{
 http://www.cpc.unc.edu/projects/addhealth/design
}
\examples{
data(YX_Friend)
}
\keyword{datasets}
