% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenmodel_setup.R
\name{eigenmodel_setup}
\alias{eigenmodel_setup}
\title{Setup constants and starting values for an eigenmodel fit}
\usage{
eigenmodel_setup(R = 0, seed = 1, em_env = .GlobalEnv)
}
\arguments{
\item{R}{non-negative integer rank of the approximating matrix}

\item{seed}{a random seed}

\item{em_env}{enviromnemt within which to do the fitting}
}
\description{
Setup constants and starting values for an eigenmodel fit
}
\author{
Peter Hoff
}
\keyword{models}
\keyword{multivariate}
