% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.eigenmodel_post.R
\name{plot.eigenmodel_post}
\alias{plot.eigenmodel_post}
\title{Plot the output of an eigenmodel fit}
\usage{
\method{plot}{eigenmodel_post}(x, ...)
}
\arguments{
\item{x}{an object of class \code{eigenmodel_post}}

\item{...}{additional plotting options}
}
\description{
A graphical display of MCMC output and posterior estimates of model
parameters in an eigenmodel fit. Includes 95 percent quantile-based
posterior confidence intervals of regression coefficients.
}
\author{
Peter Hoff
}
\keyword{multivariate}
