% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Sample from the multivariate normal distribution}
\usage{
rmvnorm(mu, Sig2)
}
\arguments{
\item{mu}{a p x 1 vector}

\item{Sig2}{a p x p positive definite matrix}
}
\value{
a p x 1 vector
}
\description{
Sample from the multivariate normal distribution
}
\examples{
rmvnorm( c(0,0,0),diag(rep(3,1)) )   
}
\author{
Peter Hoff
}
\keyword{distribution}
\keyword{models}
\keyword{multivariate}
