% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULU.R
\name{ULU}
\alias{ULU}
\title{Computes a matrix from its eigenvalue decomposition}
\usage{
ULU(UL)
}
\arguments{
\item{UL}{a list with first component ``U'', an n x r matrix and the second
component ``L'' an r x r diagonal matrix}
}
\value{
an n x n matrix
}
\description{
Computes a matrix from its eigenvalue decomposition
}
\author{
Peter Hoff
}
\keyword{models}
\keyword{multivariate}
