% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XB.R
\name{XB}
\alias{XB}
\title{Computes a sociomatrix of regression effects}
\usage{
XB(X, b)
}
\arguments{
\item{X}{an n x n x p array}

\item{b}{a p x 1 vector}
}
\value{
an n x n matrix
}
\description{
Computes a sociomatrix of regression effects
}
\author{
Peter Hoff
}
\keyword{models}
\keyword{multivariate}
