% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenmodel-package.R
\docType{data}
\name{Y_Gen}
\alias{Y_Gen}
\title{Relations between words in the 1st chapter of Genesis}
\description{
The i,j th entry of this matrix is the numerical count of the number of
times word i was next to word j in the first chapter of Genesis.
}
\examples{

data(Y_Gen)

}
\keyword{datasets}
