% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y_impute.R
\name{Y_impute}
\alias{Y_impute}
\title{Impute missing values of a sociomatrix}
\usage{
Y_impute()
}
\value{
symmetric matrix
}
\description{
Impute missing values of a sociomatrix
}
\details{
Imputes missing values of a sociomatrix from a matrix of latent variables
and an ordered-probit specification.
}
\author{
Peter Hoff
}
\keyword{models}
\keyword{multivariate}
