% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenmodel-package.R
\docType{data}
\name{Y_Pro}
\alias{Y_Pro}
\title{Butland's protein-protein interaction data}
\description{
Butland's protein-protein interaction data
}
\examples{

data(Y_Pro)

}
\references{
Butland et al (2005) ``Interaction network containing conserved
and essential protein complexes in Escherichia coli''
}
\keyword{datasets}
