% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenmodel-package.R
\docType{data}
\name{YX_Friend}
\alias{YX_Friend}
\title{Sex, race and friendship data from a 12th grade classroom}
\source{
http://www.cpc.unc.edu/projects/addhealth/design
}
\description{
A list in which \code{Y} encodes the presence of a friendship tie between 90
12th graders. The array \code{X} indicates pairs of the same sex and of the
same race.
}
\examples{

data(YX_Friend)

}
\keyword{datasets}
